api/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
pyyaml>=6,<7
anthropic>=0.49,<1
